requirements.txt
discord.py
pynacl
yt-dlp